// src/types/ICurrentWeatherData.ts
import { IWeatherData } from "./IWeatherData"

export interface ICurrentWeatherData {
  coord: {
    lon: number
    lat: number
  }
  weather: IWeatherData[]
  base: "stations"
  main: {
    temp: number
    feels_like: number
    temp_min: number
    temp_max: number
    pressure: number
    humidity: number
    sea_level: number
    grnd_level: number
  }
  visibility: number
  wind: {
    speed: number
    deg: number
    gust: number
  }
  clouds: {
    all: number
  }
  dt: number
  sys: {
    type: number
    id: number
    country: string
    sunrise: number
    sunset: number
  }
  timezone: number
  id: number
  name: string
  cod: number
}